+++ src/Component/HappyFamily/style.js
@@ -1,0 +1,8 @@
+import { styled } from '@mui/system';
+export const StyledGrid = styled('div')({
+    display: "flex",
+    flexDirection: "column",
+    justifyContent: "space-around",
+    height: 500,
+
+});
