+++ src/Component/HappyFamily/style.module.css
@@ -3,7 +3,7 @@
     border-radius: 15px 15px 15px 15px;
     margin: 15px 39px 39px 20px;
     position: absolute;
-    width: 356.13px;
-    height: 486.75px;
+    width: 357px;
+    height: 487px;
 
 }
